third_party/WebKit/LayoutTests/fast/dom/documentmarker-add-adjacent-text-expected.txt

CONSOLE ERROR: line 22: Uncaught SyntaxError: Illegal return statement
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
text
